liuxiaodong,主账号创建删除添加校验

+++ uoo-core/uoo-personnel/src/main/java/cn/ffcs/uoo/core/personnel/entity/TbAcct.java
@@ -71,7 +71,7 @@ public class TbAcct extends Model<TbAcct> {
      * 创建人
      */
     @TableField("CREATE_USER")
-    private BigDecimal createUser;
+    private Long createUser;
     /**
      * 修改时间
      */
@@ -81,7 +81,7 @@ public class TbAcct extends Model<TbAcct> {
      * 修改人
      */
     @TableField("UPDATE_USER")
-    private BigDecimal updateUser;
+    private Long updateUser;
     /**
      * 状态变更的时间
      */
@@ -160,11 +160,11 @@ public class TbAcct extends Model<TbAcct> {
         this.createDate = createDate;
     }
 
-    public BigDecimal getCreateUser() {
+    public Long getCreateUser() {
         return createUser;
     }
 
-    public void setCreateUser(BigDecimal createUser) {
+    public void setCreateUser(Long createUser) {
         this.createUser = createUser;
     }
 
@@ -176,11 +176,11 @@ public class TbAcct extends Model<TbAcct> {
         this.updateDate = updateDate;
     }
 
-    public BigDecimal getUpdateUser() {
+    public Long getUpdateUser() {
         return updateUser;
     }
 
-    public void setUpdateUser(BigDecimal updateUser) {
+    public void setUpdateUser(Long updateUser) {
         this.updateUser = updateUser;
     }
