+++ uoo-core/uoo-personnel/src/main/java/cn/ffcs/uoo/core/personnel/entity/TbAcctExt.java
@@ -86,7 +86,7 @@ public class TbAcctExt extends Model<TbAcctExt> {
      * 创建人
      */
     @TableField("CREATE_USER")
-    private BigDecimal createUser;
+    private Long createUser;
     /**
      * 修改时间
      */
@@ -96,7 +96,7 @@ public class TbAcctExt extends Model<TbAcctExt> {
      * 修改人
      */
     @TableField("UPDATE_USER")
-    private BigDecimal updateUser;
+    private Long updateUser;
     /**
      * 状态变更的时间
      */
@@ -200,11 +200,11 @@ public class TbAcctExt extends Model<TbAcctExt> {
         this.createDate = createDate;
     }
 
-    public BigDecimal getCreateUser() {
+    public Long getCreateUser() {
         return createUser;
     }
 
-    public void setCreateUser(BigDecimal createUser) {
+    public void setCreateUser(Long createUser) {
         this.createUser = createUser;
     }
 
@@ -216,11 +216,11 @@ public class TbAcctExt extends Model<TbAcctExt> {
         this.updateDate = updateDate;
     }
 
-    public BigDecimal getUpdateUser() {
+    public Long getUpdateUser() {
         return updateUser;
     }
 
-    public void setUpdateUser(BigDecimal updateUser) {
+    public void setUpdateUser(Long updateUser) {
         this.updateUser = updateUser;
     }
